Renamed ConsoleApp1 folder to JokeGenerator. Folder appears to have used default name when generated in IDE; this was changed to match the project name. Further changes may be warranted if the application is divided into multiple projects.
