+++ c-sharp/.vscode/launch.json
@@ -10,9 +10,9 @@
             "request": "launch",
             "preLaunchTask": "build",
             // If you have changed target frameworks, make sure to update the program path.
-            "program": "${workspaceFolder}/ConsoleApp1/bin/Debug/netcoreapp2.1/JokeGenerator.dll",
+            "program": "${workspaceFolder}/JokeGenerator/bin/Debug/netcoreapp2.1/JokeGenerator.dll",
             "args": [],
-            "cwd": "${workspaceFolder}/ConsoleApp1",
+            "cwd": "${workspaceFolder}/JokeGenerator",
             // For more information about the 'console' field, see https://aka.ms/VSCode-CS-LaunchJson-Console
             "console": "integratedTerminal",
             "stopAtEntry": false
