+++ c-sharp/.vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/ConsoleApp1/JokeGenerator.csproj",
+                "${workspaceFolder}/JokeGenerator/JokeGenerator.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/ConsoleApp1/JokeGenerator.csproj",
+                "${workspaceFolder}/JokeGenerator/JokeGenerator.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,11 +32,27 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/ConsoleApp1/JokeGenerator.csproj",
+                "${workspaceFolder}/JokeGenerator/JokeGenerator.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
             "problemMatcher": "$msCompile"
-        }
+        },
+        {
+            "label": "test",
+            "group": {
+              "kind": "test",
+              "isDefault": true
+            },
+            "command": "dotnet",
+            "type": "process",
+            "args": [
+              "watch",
+              "-p",
+              "${workspaceFolder}/JokeGenerator.Tests/JokeGenerator.Tests.csproj",
+              "test",
+            ],
+            "problemMatcher": "$msCompile"
+          }
     ]
 }
