+++ c-sharp/JokeGenerator.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2036
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JokeGenerator", "ConsoleApp1\JokeGenerator.csproj", "{F31733F5-8322-4DDC-86C9-658FFDC656F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JokeGenerator", "JokeGenerator\JokeGenerator.csproj", "{F31733F5-8322-4DDC-86C9-658FFDC656F8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
