+++ c-sharp/JokeGenerator/ConsolePrinter.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ConsoleApp1
+namespace JokeGenerator
 {
     public class ConsolePrinter
     {
