+++ c-sharp/JokeGenerator/JsonFeed.cs
@@ -3,7 +3,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
-namespace ConsoleApp1
+namespace JokeGenerator
 {
     class JsonFeed
     {
@@ -14,20 +14,20 @@ namespace ConsoleApp1
         {
             _url = endpoint;
         }
-        
-		public static string[] GetRandomJokes(string firstname, string lastname, string category)
-		{
-			HttpClient client = new HttpClient();
-			client.BaseAddress = new Uri(_url);
-			string url = "jokes/random";
-			if (category != null)
-			{
-				if (url.Contains('?'))
-					url += "&";
-				else url += "?";
-				url += "category=";
-				url += category;
-			}
+
+        public static string[] GetRandomJokes(string firstname, string lastname, string category)
+        {
+            HttpClient client = new HttpClient();
+            client.BaseAddress = new Uri(_url);
+            string url = "jokes/random";
+            if (category != null)
+            {
+                if (url.Contains('?'))
+                    url += "&";
+                else url += "?";
+                url += "category=";
+                url += category;
+            }
 
             string joke = Task.FromResult(client.GetStringAsync(url).Result).Result;
 
@@ -48,19 +48,19 @@ namespace ConsoleApp1
         /// <param name="client2"></param>
         /// <returns></returns>
 		public static dynamic Getnames()
-		{
-			HttpClient client = new HttpClient();
-			client.BaseAddress = new Uri(_url);
-			var result = client.GetStringAsync("").Result;
-			return JsonConvert.DeserializeObject<dynamic>(result);
-		}
+        {
+            HttpClient client = new HttpClient();
+            client.BaseAddress = new Uri(_url);
+            var result = client.GetStringAsync("").Result;
+            return JsonConvert.DeserializeObject<dynamic>(result);
+        }
 
-		public static string[] GetCategories()
-		{
-			HttpClient client = new HttpClient();
-			client.BaseAddress = new Uri(_url);
+        public static string[] GetCategories()
+        {
+            HttpClient client = new HttpClient();
+            client.BaseAddress = new Uri(_url);
 
-			return new string[] { Task.FromResult(client.GetStringAsync("categories").Result).Result };
-		}
+            return new string[] { Task.FromResult(client.GetStringAsync("categories").Result).Result };
+        }
     }
 }
