+++ c-sharp/JokeGenerator/Program.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ConsoleApp1
+namespace JokeGenerator
 {
     class Program
     {
